requirements.txt
pytelegrambotapi==4.31.0
speechrecognition==3.14.5
pydub==0.25.1
vosk==0.3.45
pyyaml==6.0.3
requests==2.32.5
pillow==12.2.0
pyzbar==0.1.9
opencv-python==4.13.0.92
sphinx==9.1.0